refactor panel to web
