--- pyhdx/panel/bokeh_extensions/index.ts
@@ -1,1 +1,0 @@
-export {ngl} from "./ngl_viewer"
