--- pyhdx/panel/config.ini
@@ -1,4 +1,0 @@
-[cluster]
-ip = 127.0.0.1
-port = 52123
-
