+++ setup.cfg
@@ -39,7 +39,7 @@ console_scripts =
 
 
 [options.extras_require]
-panel =
+web =
     panel>=0.11.0
     bokeh
     param
